+++ CHANGES.md
@@ -36,6 +36,7 @@ Changes in 12.5
 - JS remove inline code from program in Administration.php & Administration.js
 - JS remove inline code from program in FixDailyAttendance.php & FixDailyAttendance.js
 - JS remove inline code from RegistrationAdminContactEnable() in RegistrationAdmin.fnc.php & RegistrationAdminContactEnable.js
+- JS remove inline code from RegistrationSiblingUseContactsAddress() in Registration.fnc.php & RegistrationSiblingUseContactsAddress.js
 
 Changes in 12.4.3
 -----------------